Split header toast from app render
